adds declrative schema and remove scripts
