--- Setup/InstallData.php
@@ -1,99 +1,0 @@
-<?php
-/**
- * Astound
- * NOTICE OF LICENSE
- * This source file is subject to the Open Software License (OSL 3.0)
- * that is bundled with this package in the file LICENSE.txt.
- * It is also available through the world-wide-web at this URL:
- * http://opensource.org/licenses/osl-3.0.php
- * If you did not receive a copy of the license and are unable to
- * obtain it through the world-wide-web, please send an email
- * to codemaster@astoundcommerce.com so we can send you a copy immediately.
- *
- * @category  Affirm
- * @package   Astound_Affirm
- * @copyright Copyright (c) 2016 Astound, Inc. (http://www.astoundcommerce.com)
- * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
- */
-
-namespace Astound\Affirm\Setup;
-
-use Magento\Customer\Setup\CustomerSetupFactory;
-use Magento\Customer\Model\Customer;
-use Magento\Eav\Model\Entity\Attribute\Set as AttributeSet;
-use Magento\Eav\Model\Entity\Attribute\SetFactory as AttributeSetFactory;
-use Magento\Framework\Setup\InstallDataInterface;
-use Magento\Framework\Setup\ModuleContextInterface;
-use Magento\Framework\Setup\ModuleDataSetupInterface;
-
-/**
- * Install Data
- */
-class InstallData implements InstallDataInterface
-{
-    /**
-     * Customer setup factory
-     *
-     * @var CustomerSetupFactory
-     */
-    protected $customerSetupFactory;
-
-    /**
-     * @var AttributeSetFactory
-     */
-    protected $attributeSetFactory;
-
-    /**
-     * Init
-     *
-     * @param CustomerSetupFactory $customerSetupFactory
-     * @param AttributeSetFactory $attributeSetFactory
-     */
-    public function __construct(
-        CustomerSetupFactory $customerSetupFactory,
-        AttributeSetFactory $attributeSetFactory
-    ) {
-        $this->customerSetupFactory = $customerSetupFactory;
-        $this->attributeSetFactory = $attributeSetFactory;
-    }
-
-    /**
-     * {@inheritdoc}
-     * @SuppressWarnings(PHPMD.ExcessiveMethodLength)
-     */
-    public function install(ModuleDataSetupInterface $setup, ModuleContextInterface $context)
-    {
-        /** @var CustomerSetup $customerSetup */
-        $customerSetup = $this->customerSetupFactory->create(['setup' => $setup]);
-
-        $customerEntity = $customerSetup->getEavConfig()->getEntityType('customer');
-        $attributeSetId = $customerEntity->getDefaultAttributeSetId();
-
-        /** @var $attributeSet AttributeSet */
-        $attributeSet = $this->attributeSetFactory->create();
-        $attributeGroupId = $attributeSet->getDefaultGroupId($attributeSetId);
-        $attributeCode = 'affirm_customer_mfp';
-
-        $customerSetup->addAttribute(Customer::ENTITY, $attributeCode, [
-            'type'         => 'varchar',
-            'input'        => 'text',
-            'required'     => 0,
-            'label'        => 'Multiple Financing Program value',
-            'global'       => 1,
-            'visible'      => 1,
-            'user_defined' => 1,
-            'sort_order'   => 1000,
-            'position'     => 1000,
-            'system'       => 0
-        ]);
-
-        $attribute = $customerSetup->getEavConfig()->getAttribute(Customer::ENTITY, $attributeCode)
-            ->addData([
-                'attribute_set_id'   => $attributeSetId,
-                'attribute_group_id' => $attributeGroupId,
-                'used_in_forms'      => ['adminhtml_customer'],
-            ]);
-
-        $attribute->save();
-    }
-}
