--- Setup/InstallSchema.php
@@ -1,128 +1,0 @@
-<?php
-
-namespace Astound\Affirm\Setup;
-
-use Magento\Framework\Setup\InstallSchemaInterface;
-use Magento\Framework\Setup\ModuleContextInterface;
-use Magento\Framework\Setup\SchemaSetupInterface;
-
-
-class InstallSchema implements InstallSchemaInterface
-{
-    public function install(SchemaSetupInterface $setup, ModuleContextInterface $context)
-    {
-        $installer = $setup;
-        $installer->startSetup();
-
-        $table = $installer
-            ->getConnection()
-            ->newTable($installer->getTable('astound_affirm_rule'))
-            ->addColumn(
-                'rule_id',
-                \Magento\Framework\DB\Ddl\Table::TYPE_INTEGER,
-                null,
-                ['identity' => true, 'unsigned' => true, 'nullable' => false, 'primary' => true]
-            )
-            ->addColumn(
-                'for_admin',
-                \Magento\Framework\DB\Ddl\Table::TYPE_BOOLEAN,
-                null,
-                ['nullable' => false, 'unsigned' => true, 'default' => 0]
-            )
-            ->addColumn(
-                'is_active',
-                \Magento\Framework\DB\Ddl\Table::TYPE_BOOLEAN,
-                null,
-                ['nullable' => false, 'unsigned' => true, 'default' => 0]
-            )
-            ->addColumn(
-                'all_stores',
-                \Magento\Framework\DB\Ddl\Table::TYPE_BOOLEAN,
-                null,
-                ['nullable' => false, 'unsigned' => true, 'default' => 0]
-            )
-            ->addColumn(
-                'all_groups',
-                \Magento\Framework\DB\Ddl\Table::TYPE_BOOLEAN,
-                null,
-                ['nullable' => false, 'unsigned' => true, 'default' => 0]
-            )
-            ->addColumn(
-                'name',
-                \Magento\Framework\DB\Ddl\Table::TYPE_TEXT,
-                255,
-                ['default' => null, 'nullable' => false]
-            )
-            ->addColumn(
-                'stores',
-                \Magento\Framework\DB\Ddl\Table::TYPE_TEXT,
-                255,
-                ['default' => '', 'nullable' => false]
-            )
-            ->addColumn(
-                'cust_groups',
-                \Magento\Framework\DB\Ddl\Table::TYPE_TEXT,
-                255,
-                ['default' => '', 'nullable' => false]
-            )
-            ->addColumn(
-                'message',
-                \Magento\Framework\DB\Ddl\Table::TYPE_TEXT,
-                null,
-                ['default' => '', 'nullable' => false]
-            )
-            ->addColumn(
-                'methods',
-                \Magento\Framework\DB\Ddl\Table::TYPE_TEXT,
-                null,
-                ['default' => null, 'nullable' => true]
-            )
-            ->addColumn(
-                'conditions_serialized',
-                \Magento\Framework\DB\Ddl\Table::TYPE_TEXT,
-                null,
-                ['default' => null, 'nullable' => true]
-            );
-
-        $installer->getConnection()->createTable($table);
-
-        $table = $installer
-            ->getConnection()
-            ->newTable($installer->getTable('astound_affirm_attribute'))
-            ->addColumn(
-                'attr_id',
-                \Magento\Framework\DB\Ddl\Table::TYPE_INTEGER,
-                null,
-                ['identity' => true, 'unsigned' => true, 'nullable' => false, 'primary' => true]
-            )
-            ->addColumn(
-                'rule_id',
-                \Magento\Framework\DB\Ddl\Table::TYPE_INTEGER,
-                null,
-                ['unsigned' => true, 'nullable' => false]
-            )
-            ->addColumn(
-                'code',
-                \Magento\Framework\DB\Ddl\Table::TYPE_TEXT,
-                255,
-                ['default' => null, 'nullable' => true]
-            )
-            ->addIndex('rule_id', 'rule_id')
-            ->addForeignKey(
-                $installer->getFkName(
-                    'astound_affirm_attribute',
-                    'rule_id',
-                    'astound_affirm_rule',
-                    'rule_id'
-                ),
-                'rule_id',
-                $installer->getTable('astound_affirm_rule'),
-                'rule_id',
-                \Magento\Framework\DB\Ddl\Table::ACTION_CASCADE
-            );
-
-        $installer->getConnection()->createTable($table);
-
-        $installer->endSetup();
-    }
-}
