+++ Setup/db_schema.xml
@@ -1,0 +1,28 @@
+<?xml version="1.0"?>
+<schema xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:Setup/Declaration/Schema/etc/schema.xsd">
+  <table name="astound_affirm_rule" resource="default" engine="innodb" comment="Affirm Rule Table">
+    <column xsi:type="int" name="rule_id" unsigned="true" nullable="false" identity="true" comment="Rule ID"/>
+    <column xsi:type="boolean" name="for_admin" nullable="false" default="0" comment="For Admin"/>
+    <column xsi:type="boolean" name="is_active" nullable="false" default="0" comment="Is Active"/>
+    <column xsi:type="boolean" name="all_stores" nullable="false" default="0" comment="All Stores"/>
+    <column xsi:type="boolean" name="all_groups" nullable="false" default="0" comment="All Groups"/>
+    <column xsi:type="varchar" name="name" length="255" nullable="false" comment="Name"/>
+    <column xsi:type="varchar" name="stores" length="255" nullable="false" comment="Stores"/>
+    <column xsi:type="varchar" name="cust_groups" length="255" nullable="false" comment="Customer Groups"/>
+    <column xsi:type="text" name="message" nullable="false" comment="Message"/>
+    <column xsi:type="text" name="methods" nullable="true" comment="Methods"/>
+    <column xsi:type="text" name="conditions_serialized" nullable="true" comment="Serialized Conditions"/>
+    <constraint xsi:type="primary" referenceId="PRIMARY">
+      <column name="rule_id"/>
+    </constraint>
+  </table>
+  <table name="astound_affirm_attribute" resource="default" engine="innodb" comment="Affirm Attribute Table">
+    <column xsi:type="int" name="attr_id" unsigned="true" nullable="false" identity="true" comment="Attribute ID"/>
+    <column xsi:type="int" name="rule_id" unsigned="true" nullable="false" identity="false" comment="Rule ID"/>
+    <column xsi:type="varchar" name="code" length="255" nullable="true" comment="Code"/>
+    <constraint xsi:type="primary" referenceId="PRIMARY">
+      <column name="attr_id"/>
+    </constraint>
+    <constraint xsi:type="foreign" referenceId="AFFIRM_ATTR_RULE_ID_RULE_RULE_ID" table="astound_affirm_attribute" column="rule_id" referenceTable="astound_affirm_rule" referenceColumn="rule_id" onDelete="CASCADE"/>
+  </table>
+</schema>
